requirements.txt
asgiref==3.7.2
Django==4.2.11
djangorestframework==3.15.2
gyp-next==0.17.0
iniconfig==2.1.0
numpy==2.3.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytz==2024.1
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
django-bootstrap5
